update: retreive items from db
